Big refactor: Py3, --user installations, pipenv and pipx
The install script now creates a virtualenv (via pipenv (via pipx)) so
the .py scripts can run when another virtualenv is active. Previously,
the scripts just tried to run with the current python interpreter which
would fail in a Py3 venv (up to now scripts were Py2) or if libraries
were mismatched.

The minimum Python version is now 3.8, as that's the Python3 version
that comes with macOS 11.0 (Big Sur)

There is now a bin directory, like many libraries that install
executables. The contents are gitignored, and can be discarded and
regenerated based on a manifest. This lets us use the bin folder as a
convenient dumping ground for anything we want available on the $PATH,
not just scripts from this repo. One new use is to symlink first-party
command-line launchers from utilities I use. This allows for their use
without admin priviliges, which is useful for my new work laptop. Where
there are no first-party launchers, open -b bins are still created.

Other bits and bobs:
- Removed: newscript
  Fell out of use. I either type the interpreter name, add it to a task
  runner, or put it in this repo, depending on the pattern of use
- Removed: pylint2/3
  I do this through pipenv these days
- Removed: search-bitbucket
  yeah
- Removed: venv-reboot
  pipenv again
- Removed: screensaver
  never used it
- Removed: vba
  Switched to pipenv
- finder-(show|hide)-hidden is now a regular script - it's not
  frequently used - it feels better as a script rather than polluting
  the function and alias namespace. Also, `killall Finder` seemed a bit
  sledgehammery; now it takes a bit more care ensuring it's terminating
  the right task